REEB-55
git-svn-id: http://svn.uhuila.cn/reeb/trunk@231 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ module/accounts/app/models/accounts/Accounts.java
@@ -1,6 +1,7 @@
 package models.accounts;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 import play.db.jpa.Model;
 
@@ -9,10 +10,10 @@ import play.db.jpa.Model;
 public class Accounts extends Model {
     public long uid;
     public String accountType;
-    public float totalAmount;
+    public BigDecimal totalAmount;
 
 
-    public Accounts(long uid, String accountType, float totalAmount){
+    public Accounts(long uid, String accountType, BigDecimal totalAmount){
         this.uid = uid;
         this.accountType = accountType;
         this.totalAmount = totalAmount;
